UnitController is maybe about 50% done!
-Added a deepCloning method to UnitController;
-NOTE: Had to made Unit, Character, and Job Serializable to facilitate a
deep clone.
-Worked on UnitController. UnitController can now build a local sheet,
which just creates a character from scratch at their base level, and
levels them up based on the classHistory.
-Note, there will be two sheets stored in UnitController at any time.
One which generates a unit for comparison ,this is the local one. And
another that generates the sheet based on input. This is the input one.
-Tested like all the methods in UnitController. Everything works so far.
-Reclassing works. Check out Silas's stint as a Songstress :O
-TODO: Nothing checks for max stats yet, need to implement this. Also
will need to format the decimals at some point ,they're pretty ugly.
Also will have to make UnitController a singleton at some point...