Calculate Button now works! Also some big UI Changes...
-- The Level field in the main Window, Graph panel, and Reclass window
are all now combo boxes. The main panel one is tied to the character
now, so selecting a character will limit your options based on the base
level of that character. Also, selecting a level in the main window will
change the boxes in the graph menu...

-- The Calculate Button now works! Had to make a few changes to
UnitController, both to fix mistakes and also to reformat some things.
UnitController now has a method to return the class history as an String
Array with the levels properly appended (Previously this was handled in
the GUI and added to the Controller afterwards, which messes up the code
for building a sheet).

-Next up! The graph :O