Starting the Graph!
-So first, for the graph component we'll be using the JFreeChart
library. The necessary JAR files are included in the external libraries
folder.

-Added a graph controller. This graph controller creates the graph
component at the GUI, and is used for updating it whenever the
parameters are changed .This will occur when the calculate button is
clicked (generating the all the data and the graph), as well as when any
of the graph options are changed (Switching between stats).

-For now, the graph appears by default, and the calculate button will
just bring up some random lines just for testing purposes.

-Removed some random placeholders classes that hadn't been removed yet!