Graph component and Calculate component now fully implemented
-The graph now properly displays when the calculate button is clicked.
Additionally, all the values in the results field now populate

-I added a third box to the results field to show your  expected values
next to the general expected values, and a level slider to adjust the
level. Not sure how to communicate to the user which box is their
characters expected output, and which is the general expected output,
but ill figure that out later.

-Numbers are formatted and the difference box changes color if your...
blessed or screwed :D