Child Fixes
-Resized parentalUnitsPane to be wider as things were being clipped
depending on the variedParent and class
-Created a method for calculating the base level as it occurred several
times in our code
-Created method for clearing the results boxes
-Selecting a new character now clears the boxes
-Added a check in calculate saying that if parents weren't set up, that
the user should set them up. This way calculate doesn't print out the
data of the previously selected character
-Changing to every route now automatically selects silas (for debugging)
-Made it so the pop up when you select a child to input parent stats now
appears over the GUI rather than the center of the user's screen
-Fixed issue where children were not starting at the correct base class
and the sheets were being built wrongly (problem in
CalculateButtonHandler)
-Added functionality to let the user change a Child Character's starting
level and retain any reclassing they had done - this, however, after
thinking about it, is probably unnecessary functionlity. T.T It's been
left there, commented out, because it was hard to figure out and cool.
>:(  (also we may decided to use it later)