Updating Reclasing and Adding Inner Level Concept
-Added the concept of inner levels throughout all of the code. For
purposes of code, all units maintain an inner level and change to show
the appropriate display level. In other words, a level 5 Paladin
promoted at level 20 from Cavalier will have an inner level of 35 and a
display level of 5. This assists in reading the graph as well as for
reclassing purposes
-Added constants in DataStorage for the base max level of 20 and the
special max level of 40
-Prepromoted units add 20 to their starting level for the purposes of
inner levels. This is done when creating the character in CharBoxHandler
through the variable prepromoteModifier (NOTE: This is not done with
children as children will never be prepromoted)
-Adjusted ClearButtonHandler to account for inner levels
-Modified and methodized setting possible levels in any place that may
end up dealing with promoted units resulting in inner and display levels
-Removed calculateBaseLevel as the concept of inner levels invalidated
it. Wherever the method was called now retrieves the StartLevel from
unitcontroller as that has replaced baseLevel in the GUI
-Move Child Options Window listeners to their own block divided from the
Option Window block. It now has it's own comment header
-Reclassing now reclasses all levels until the current class in the
classHistory isn't the same as the class that the program began
reclassing from. In other words, it will not overwrite reclasses made
further in the classHistory by the user
-changed getClassArray(int) in UnitController to
getFormattedClassHistory() - the parameter was removed as
-getFormattedClassHistory now accounts for inner levels when forming
it's history.
-Changed how promotedLevel is calculate in getFormattedClassHistory to
account for prepromoted units
-Changed some ActionListener Class names to bring them to standard
conventions (some of them did not have their first letter capitalized)